Did a first thorough revision of the unit tests. Not all are completed yet. Fixed some bugs as well: Quality wasn't properly updated in the Solvent and CannaOil classes. This was due to the Solvent constructor containing a Quality = 0 sentence. Now it properly updates from solventmix.chronic.
Also tests are now named differently to make more clear what they are and what they do. The first part of the sentence is to which gameplay element the classes belong, then the action tested in the element, and  the expected result. Ex.: CannaOilCreation_CreateMasterKushSolvent_TypeofSolvent.

I'll work more on later tonight or tomorrow.